♻️ Refactor to use more ts
