--- template/packages/localization/config/index.ts
@@ -1,1 +1,0 @@
-export { defaultLocale, languages, locales } from './langs.cjs';
