+++ template/packages/localization/index.ts
@@ -1,3 +1,3 @@
-export * from './config';
 export * from './translations';
 export * from './types';
+export * from './lokse.config';
