+++ template/packages/localization/package.json
